Add files via upload
Three Jesuits and three cannibals must cross a river. However they only have one boat
with capacity for two people. The security measure does not allow some of the
the river banks with a number of Jesuits are less than the number of cannibals. What is the sequence of
necessary trips to cross the river safely for the Jesuits?